refactor(datepicker): uncircularize dependencies
